feat: split combox field into control and field composables
