+++ packages/core/src/useComboBox/index.ts
@@ -1,1 +1,2 @@
 export * from './useComboBox';
+export * from './useComboxBoxControl';
